requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==21.2.0
black==19.10b0
Brotli==1.0.9
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
httprunner==3.1.6
idna==3.2
importlib-metadata==4.6.1
Jinja2==2.11.3
jmespath==0.9.5
loguru==0.4.1
MarkupSafe==2.0.1
more-itertools==8.8.0
packaging==21.0
pathspec==0.9.0
pluggy==0.13.1
py==1.10.0
pydantic==1.8.2
pyparsing==2.4.7
pytest==5.4.3
pytest-html==2.1.1
pytest-metadata==1.11.0
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
sentry-sdk==0.14.4
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
win32-setctime==1.0.3
zipp==3.5.0
